Adjusted trace output to comply with sequence.
I wrote a sequence.rb program a long time ago that would write sequence
diagrams for me.  It consumes trace output and turns them into sequence
diagrams.  For it to work it needed the trace to have a particular form.

This could be adjusted and improved, but for now it is good enough to
get sequence diagrams into the documentation.