Moving event serialization/de-serialization code
This was done to take complexity away from the user of the networking
classes.  The Event object is dynamic so it can't be pickled.  For this
reason it has it's own serialization and de-serialization code.  It is
not reasonable to ask a client to remember to use this code prior to
sending and receiving a message over a network.

Also, it would spread functionality that should be grouped in one
location across the code base, making the code harder to maintain.