Fixed an issue with the live trace.
In a previous version, the live trace would output the last observed
trace no matter if a state transition was caused by the triggering
signal.  This is exactly how we don't want the trace to work.  Instead
of reproducing how the trace log was being build up in the printing
code, we instead just look at the length of this trace log.  If it has
changed from the last time we print... otherwise, we don't.