Adjusted to exclude local none-connected ip addresses.
The rabbitmq server connects to all ip addresses on a machine.  This can
be adjusted by changing the RabbitMq configuration file; but I do not
want to touch it because it's example config is broken out of the box.
There are a lot of things broken with RabbitMq and the documentation is
extensive and illegible.

The code as it was currently written transmits messages to the windows
machine twice.  The update, only calls one connection for the local
machine. This fixes the issue of unwanted multiple messaging.