Making library resilient to stop/start actions.
In the case that a Factory or ActiveObject is stopped, and items are
sent to it, using a one-shot or a multishot, a race condition can occur
between code trying to read the queue and the code trying to remove the
queue.  To address this, the len of the queue checked prior to accessing
it.