Perf: If instrumentation is False, spy_on skipped
The user should have the ability to turn off instrumentation without
having to go through their code to remove all of their spy_on
decorators.  Now, if the instrumentated attribute is set to False, the
spy_on decorator stubs itself, almost behaving as it is were not in the
code at all.