Adding a pytest.ini file to remove NEW pytest warnings.
After updating the pytest package, and running the tests, I saw many
many warning messages that weren't there before.  It seems that they
have added a "feature" which requires you to place a pytest.ini file
listing all of your markers.

So now the markers are defined in two locations.  In the tests and in an
ini file.  This is bad design pytest.