Working on thread safe attributes.
A set of thread safe attributes can be constructed by making a list of
names in your class scope named '_attributes' if you inherit from the
ThreadSafeAttributes class.  The ThreadSafeAttributes class uses the
MetaThreadSafeAttributes class, which constructs a set of
ThreadSafeAttribute objects, which use the descriptor protocol to wrap a
deque containing one location.