Unproved fix for thread safe attributes.
I can not get thread-unsafe-code to fail any of my tests in the WSL.  I
will run this same code on a linux server when I get access to one.
Also, the valgrind tool helgrind, issued failures from a simple
multithreaded python example.  Once again, this could be an issue with
the WSL, and/or valgrind and python my not play well together.