Adjustment to example to make "coverage" work.
This change was made in response to the bug raised by "TrailingDots".
Some test code in the examples folder was interfering with how 'pytest'
and 'coverage' were evaluating the tests in the 'tests' folder.

"TrailingDots" recommended a change and I made the change so that
his/her tests could pass.